fix: refactor BatchOperationWriter classes to avoid duplications
* this will fix the bug with process/desicion empty names for Opensearch,
as Elasticsearch implementation was fixed before and this is the one that
was pulled to AbstractBatchOperationWriter
* it looks like we also had a bug with in multi-tenancy part of code,
as in Opensearch code the checks for tenant were not present

closes #26658
